backend/users/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
pillow
python-dotenv
django-simple-history
drf-yasg
requests